Compact path
When multiple belong_to-ses exist with shallow path the path needs to be
compacted to support one or more of the parameters not being supplied
whilst maintaining the correct URL when path is concatenated with
`site`.

e.g.

```
class Something
  belongs_to :foo, shallow_path: true
  belongs_to :bar, shallow_path: true
end

Something.where(bar: 2) -> "bar/2/somethings"
```

Prior to this fix the generated path was:

```
Something.where(bar: 2) -> "/bar/2/somethings"
```

The leading `/` caused issues.